refactor: move models folder to db folder
