+++ src/app/db/repository/PessoaRepository.js
@@ -1,4 +1,4 @@
-const { Usuario, Pessoa } = require('../../../../models')
+const { Usuario, Pessoa } = require('../models')
 
 var PessoaRepository = function (app) { }
