+++ src/app/db/repository/ProdutoRepository.js
@@ -1,4 +1,4 @@
-const { Produto } = require('../../../models')
+const { Produto } = require('../models')
 
 var ProdutoRepository = function (app) { }
